Huge improvement on model validation process. This improvement is related mainly to reverse engineered models that, in theory, aren't invalid but in previous versions pgModeler was raising validation errors and the user needed to fix those problems even the model being consistent on database server. Now pgModeler honors the imported structure and in a few cases of db import there will be inconsistencies.
The objects rule, index and trigger will have the SQL/XML code generated
outside of table's definition due to validation module that sometimes
needs to swap id's of those objects and tables.

The swap object ids dialog was improved. Added a button to swap the
values of the fields.

The object selector will trigger the object selection if user click the
input field

Fixed the import of rule object on DatabaseImportHelper

The model objects widget gained a filter field that is capable to filter
objects by their name or by their id

Fixed a bug when displaying the cardinalities for fk relationships
(issue #512)