Schemas by default will be created to show the bouding rectangle.
Added the method Catalog::getObjectsOIDs in order to retrieve all
database objects oid and store them in maps. This method is used as an
auxiliary for model-database diff process.

Minor improvement on BaseObject::getTypes() in order to exclude some
types from the resulting vector.

Added the method ConnectionsConfigWidget::fillConnectionsComboBox() in
order to reuse the code that fills up a combobox with configured
connections.

ModelDatabaseDiffForm implementation checkpoint.