Methods setIgnoreEmptyAttributes and setIgnoreUnkownAttributes at SchemaParser where renamed to ignoreEmptyAttributes and ignoreUnkownAttributes
Added the method SchemaParser::extractAttributes that is capable to
extract the attributes used in a loaded buffer.

Minor improvements on ModelNavigationWidget to show the filename
associated to a hovered item on combobox.

Added the method SnippetsConfigWidget::getParsedSnippet in order to
return the processed code of a snippet if this one can be parsed through
SchemaParser instance.