Minor fixes in the *::getAlterDefinition methods do avoid crashes due to null objects handling.
Fixed a crash when generating recursive views code.

Additional improvement in diff process to avoid generate code for an
unmodifiable object when its code don't differ from the same object in
database.

Added an option to DatabaseImportHelper to avoid the fk relationship
updates. This will reduce the time to perform the import step in diff.