Removed the ReadOnlyItemDelegate
Created the PlainTextItemDelegate replacing the ReadOnlyItemDelegate
when needed

Minor adjustments in queries generated in DataManipulationForm in order
to use PostgreSQL like escaping E'' permitting user to use special char
in the middle of values.

Simple quotes (') in DataManipulationForm and TableDataWidget will be
automatically replaced by double quotes ('') in order to avoid broken
commands.