Added support to use middle button to handle panning mode (#782)
Added a more user friendly message at startup whenever a missing or
corrupted configuration file is detected. The user is now presented to a
an option to restore default settings for the problematic file.

Now any default file restored in ConfigurationForm has a backup saved
into the directory backups inside the configuration storage.

Minor improvements in GeneralConfigWidget. Added an readonly input that
exposes the path to the current user's configuration storage.