Fix resolving PGMODELER_ROOT from pgmodeler.vars
When pgmodeler.vars is source'd the $0 variable expands to '-bash'
which causes readlink to fail and PGMODELER_ROOT being set to the wrong
directory.
This is fixed by using a while loop that recursively changes directory
following symlinks along the way util it reaches the end and then reads
the current working directory.
For more info, see:
https://gist.github.com/TheMengzor/968e5ea87e99d9c41782