Adding the ability to view references to store referenced tables
This feature will cause relationships to be created between the view and
the referenced tables. This is useful when we're using reverse
engineering feature in which, in previous versions, couldn't determine
the tables that were linked to a view. Now, with this commit, a
relationship is created between the view and all involved tables.

Pending only the UI to allow the user to specify its desired referenced
tables on ReferenceWidget