(WIP) Adding support to move objects on the Z stack up and down.
This is based upon the PR #1395, but with a difference that the Z move
is persisted per object in the dbm file. Also, the Z move is available
for tables, view, foreign tables and textboxes. Relationshps and schemas
are not moveable in Z stack.

Pending the creation of a proper popup menu entry for the action (as
well in other toolbars if needed)