Reverting the changes of the commit dd0dcbda since it was breaking the importing of user-defined types that wasn't related to extensions.
Changed the behavior of extension types registration, now they will be
forcibly registered with complete name (schema.name) in order to follow
the same rule of user-defined data types so the reverse engineering can
work properly for both types handled by extensions and those not.

Improved the type searching on PgSqlTypeWidget in such a way to display
a popup with the data types matching the text typed by the user.